+++ Sandbox/TestRepositoryGeneration/DataObjects/BaseRepositories/Address.cs
@@ -10,7 +10,8 @@ namespace TestRepositoryGeneration.DataObjects.BaseRepositories
     ///     Repository interface <see cref="IAddressRepository"/>
     ///     Repository <see cref="AddressRepository"/>
     /// </summary>
-    [DataAccess(TableName = "dbo.Addresses", PostgresTableName = "archive.addresses", FilterKey1 = "Country", FilterKey2 = "Country,City", FilterKey3 = "Country,City,ZipCode", IsDeleted = false)]
+    [DataAccess(TableName = "dbo.Addresses", FilterKey1 = "Country", FilterKey2 = "Country,City", FilterKey3 = "Country,City,ZipCode", IsDeleted = false)]
+    [DataArchive(TableName = "archive.addresses", FilterKey1 = "Country", FilterKey2 = "Country,City", FilterKey3 = "Country,City,ZipCode", IsDeleted = false)]
     public class Address : ITenantUnrelated
     {
         [Key]
