+++ Sandbox/TestRepositoryGeneration/DataObjects/BaseRepositories/CustomerSubscription.cs
@@ -9,7 +9,8 @@ namespace TestRepositoryGeneration.DataObjects.BaseRepositories
     ///     Repository interface <see cref="ICustomerSubscriptionRepository"/>
     ///     Repository <see cref="CustomerSubscriptionRepository"/>
     /// </summary>
-    [DataAccess(TableName = "CustomerSubscriptions", PostgresTableName = "archive.customer_subscriptions")]
+    [DataAccess(TableName = "CustomerSubscriptions")]
+    [DataArchive(TableName = "archive.customer_subscriptions")]
     public class CustomerSubscription
     {
         [Key]
