+++ Sandbox/TestRepositoryGeneration/DataObjects/BaseRepositories/ElectronicCouponsTyped.cs
@@ -10,7 +10,8 @@ namespace TestRepositoryGeneration.DataObjects.BaseRepositories
     ///     Repository interface <see cref="IElectronicCouponsTypedRepository"/>
     ///     Repository <see cref="ElectronicCouponsTypedRepository"/>
     /// </summary>
-    [DataRepository(TableName = "ElectronicCouponsTyped", IsDeleted = false, Identity = false, PostgresTableName = "archive.electronic_coupons_typed")]
+    [DataRepository(TableName = "ElectronicCouponsTyped", IsDeleted = false, Identity = false)]
+    [DataArchive(TableName = "archive.electronic_coupons_typed", IsDeleted = false, Identity = false)]
     public class ElectronicCouponsTyped : ElectronicCoupon
     {
         [Key]
