+++ Sandbox/TestRepositoryGeneration/Infrastructure/DataAccessAttribute.cs
@@ -9,8 +9,6 @@ namespace TestRepositoryGeneration.Infrastructure
 
         public string TableName { get; set; }
 
-        public string PostgresTableName { get; set; }
-
         public string TableVersion { get; set; }
 
         public string FilterKey1 { get; set; }
@@ -30,7 +28,23 @@ namespace TestRepositoryGeneration.Infrastructure
 
         public string TableName { get; set; }
 
-        public string PostgresTableName { get; set; }
+        public string TableVersion { get; set; }
+
+        public string FilterKey1 { get; set; }
+
+        public string FilterKey2 { get; set; }
+
+        public string FilterKey3 { get; set; }
+
+        public bool IsDeleted { get; set; }
+    }
+
+    [AttributeUsage(AttributeTargets.Class, AllowMultiple = true)]
+    public sealed class DataArchiveAttribute : Attribute
+    {
+        public bool Identity { get; set; }
+
+        public string TableName { get; set; }
 
         public string TableVersion { get; set; }
