+++ Sandbox/TestRepositoryGeneration/RepositoryInterfaces/IAddressRepository.cs
@@ -15,6 +15,7 @@ namespace TestRepositoryGeneration.RepositoryInterfaces
         Task<Address> GetByCountryAndCityAndZipCodeAsync(string country, string city, string zipCode, bool? isDeleted);
         Task UpdateByIdAsync(Address address);
         Task RemoveByIdAsync(Address address);
+        Task InsertOrUpdateAsync(Address address);
     }
 
 }
