+++ WCF-Generator.exe.config
@@ -29,7 +29,7 @@
             RepositoryClassProjects="TestRepositoryGeneration"
             AdditionalProjects=""
             DefaultNamespace="TestRepositoryGeneration"
-            IsPostgresDb="true"/>
+            DatabaseType="1"/>
 
       <add Name ="Sandbox.FrontEnd"
           TargetProjectName="TestRepositoryGeneration"
@@ -40,7 +40,8 @@
           RepositoryInterfacesProjectName="TestRepositoryGeneration"
           RepositoryClassProjects="TestRepositoryGeneration"
           AdditionalProjects=""
-          DefaultNamespace="TestRepositoryGeneration"/>
+          DefaultNamespace="TestRepositoryGeneration"
+          DatabaseType="1"/>
       
     </repositoryProjects>
   </repositoryGenerator>
