+++ WCF-Generator/App.config
@@ -67,7 +67,8 @@
           RepositoryInterfacesProjectName="TestRepositoryGeneration"
           RepositoryClassProjects="TestRepositoryGeneration"
           AdditionalProjects="" 
-          DefaultNamespace="TestRepositoryGeneration"/>
+          DefaultNamespace="TestRepositoryGeneration"
+          DatabaseType="1"/>
 
       <add Name ="Sandbox.FrontEnd"
           TargetProjectName="TestRepositoryGeneration"
@@ -78,11 +79,12 @@
           RepositoryInterfacesProjectName="TestRepositoryGeneration"
           RepositoryClassProjects="TestRepositoryGeneration"
           AdditionalProjects="" 
-          DefaultNamespace="TestRepositoryGeneration"/>
+          DefaultNamespace="TestRepositoryGeneration"
+          DatabaseType="1"/>
 
       <add Name ="Archive"
           TargetProjectName="TestRepositoryGeneration"
-          RepositoryAttributeName="DataAccess"
+          RepositoryAttributeName="DataArchive"
           RepositoryBase ="TestRepositoryGeneration.Infrastructure.RepositoryBase"
           RepositoryTargetFolder="PostgresRepositories"
           RepositorySuffix="ArchiveRepository"
@@ -90,7 +92,7 @@
           RepositoryClassProjects="TestRepositoryGeneration"
           AdditionalProjects=""
           DefaultNamespace="TestRepositoryGeneration"
-          IsPostgresDb="true"/>
+          DatabaseType="2"/>
       
     </repositoryProjects>
   </repositoryGenerator>
