+++ WCF-Generator/RepositoriesGeneration/Analysis/SolutionSyntaxWalker.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using WCFGenerator.RepositoriesGeneration.Heplers;
+using WCFGenerator.RepositoriesGeneration.Helpers;
 
 namespace WCFGenerator.RepositoriesGeneration.Analysis
 {
