+++ WCF-Generator/RepositoriesGeneration/Enums/DatabaseType.cs
@@ -1,0 +1,8 @@
+﻿namespace WCFGenerator.RepositoriesGeneration.Enums
+{
+    public enum DatabaseType
+    {
+        MSSql = 1,
+        PostgreSql = 2
+    }
+}
