+++ WCF-Generator/RepositoriesGeneration/Services/RepositoryService.cs
@@ -1,13 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using WCFGenerator.Common;
 using WCFGenerator.RepositoriesGeneration.Analysis;
 using WCFGenerator.RepositoriesGeneration.Configuration;
 using WCFGenerator.RepositoriesGeneration.Core;
-using WCFGenerator.RepositoriesGeneration.Heplers;
+using WCFGenerator.RepositoriesGeneration.Enums;
+using WCFGenerator.RepositoriesGeneration.Helpers;
 using WCFGenerator.RepositoriesGeneration.Infrastructure;
 using WCFGenerator.RepositoriesGeneration.Yumapos.Infrastructure.Clone.Attributes;
 using MethodInfo = WCFGenerator.RepositoriesGeneration.Infrastructure.MethodInfo;
@@ -138,21 +138,43 @@ namespace WCFGenerator.RepositoriesGeneration.Services
 
                 #endregion
 
-                var list = new List<BaseCodeClassGeneratorRepository>
+                var list = new List<RepositoryCodeGeneratorAbstract>();
+
+                // select generators by database type
+                if (r.RepositoryInfo.DatabaseType == DatabaseType.MSSql)
+                {
+                    list = new List<RepositoryCodeGeneratorAbstract>
                 {
                     // version repository
-                    new CodeClassGeneratorVersionsRepository {RepositoryInfo = r.RepositoryInfo},
+                    new VersionsRepositoryCodeGenerator {RepositoryInfo = r.RepositoryInfo},
                     // cache repository
-                    new CodeClassGeneratorCacheRepository {RepositoryInfo = r.RepositoryInfo}
+                    new CacheRepositoryCodeGenerator {RepositoryInfo = r.RepositoryInfo}
                 };
-                // Skip service-repository for many2many model
-                if (!r.RepositoryInfo.IsManyToMany)
-                {
-                    list.Add(new CodeClassGeneratorVersionedRepositoryService { RepositoryInfo = r.RepositoryInfo });
+                    // Skip service-repository for many2many model
+                    if (!r.RepositoryInfo.IsManyToMany)
+                    {
+                        list.Add(new CodeClassGeneratorVersionedRepositoryService { RepositoryInfo = r.RepositoryInfo });
+                    }
                 }
+                else if (r.RepositoryInfo.DatabaseType == DatabaseType.PostgreSql)
+                {
+                    list = new List<RepositoryCodeGeneratorAbstract>
+                {
+                    // version repository
+                    new VersionsRepositoryCodeGenerator {RepositoryInfo = r.RepositoryInfo},
+                    // cache repository
+                    new CacheRepositoryCodeGenerator {RepositoryInfo = r.RepositoryInfo}
+                };
+                    // Skip service-repository for many2many model
+                    if (!r.RepositoryInfo.IsManyToMany)
+                    {
+                        list.Add(new CodeClassGeneratorVersionedRepositoryService { RepositoryInfo = r.RepositoryInfo });
+                    }
+                }
+
                 return list;
             });
-            
+
             canBeGeneratedRepositories = canBeGeneratedRepositories.Where(r => !r.RepositoryInfo.IsVersioning).Concat(versioned).ToList();
 
             // Skiped with error
@@ -169,7 +191,7 @@ namespace WCFGenerator.RepositoriesGeneration.Services
         ///     Get general info about candidate for repository
         /// </summary>
         /// <param name="doClass">Class syntax</param>
-        private BaseCodeClassGeneratorRepository GetRepository(ClassDeclarationSyntax doClass)
+        private RepositoryCodeGeneratorAbstract GetRepository(ClassDeclarationSyntax doClass)
         {
             var className = doClass.Identifier.Text;
 
@@ -204,9 +226,7 @@ namespace WCFGenerator.RepositoriesGeneration.Services
             var dataAccess = (DataAccessAttribute)dataAccessAttr;
 
             var tableName = dataAccess.TableName ?? doClass.Identifier.ToString() + 's';
-            var postgresTableName = dataAccess.PostgresTableName ?? doClass.Identifier.ToString() + 's';
             repositoryInfo.TableName = tableName;
-            repositoryInfo.PostgresTableName = postgresTableName;
 
             // Get filters
             var filterKeys = (new[] { dataAccess.FilterKey1, dataAccess.FilterKey2, dataAccess.FilterKey3 })
@@ -393,14 +413,15 @@ namespace WCFGenerator.RepositoriesGeneration.Services
                 repositoryInfo.RequiredNamespaces.Add(RepositoryType.Cache, new List<string> { repositoryInterfaceNamespace });
             }
 
-            repositoryInfo.IsPostgresDb = _config.IsPostgresDb;
+            repositoryInfo.DatabaseType = (DatabaseType)_config.DatabaseType;
 
             #endregion
 
-            var repositoryAndDo = new CodeClassGeneratorRepository
+            var repositoryAndDo = new RepositoryCodeGenerator
             {
-                RepositoryInfo = repositoryInfo,
+                RepositoryInfo = repositoryInfo
             };
+            
             // return general repository
             return repositoryAndDo;
         }
@@ -419,8 +440,9 @@ namespace WCFGenerator.RepositoriesGeneration.Services
             // Add common methods
             var methods = new List<MethodImplementationInfo>
             {
-                new MethodImplementationInfo() { Method = RepositoryMethod.GetAll},
-                new MethodImplementationInfo() { Method = RepositoryMethod.Insert}
+                new MethodImplementationInfo { Method = RepositoryMethod.GetAll },
+                new MethodImplementationInfo { Method = RepositoryMethod.Insert },
+                new MethodImplementationInfo { Method = RepositoryMethod.InsertOrUpdate}
             };
 
             // Methods by keys from model (without methods from base model)
