+++ WCF-Generator/RepositoriesGeneration/Yumapos.Infrastructure.Clone/Attributes/DataAccessAttribute.cs
@@ -7,8 +7,6 @@ namespace WCFGenerator.RepositoriesGeneration.Yumapos.Infrastructure.Clone.Attri
     {
         public string TableName { get; set; }
 
-        public string PostgresTableName { get; set; }
-
         public string TableVersion { get; set; }
 
         public string FilterKey1 { get; set; }
