+++ WCF-Generator/WCF-Generator.csproj
@@ -172,23 +172,25 @@
     <Compile Include="DecoratorGeneration\Infrastructure\DecoratorInfo.cs" />
     <Compile Include="RepositoriesGeneration\Analysis\SolutionSyntaxWalker.cs" />
     <Compile Include="RepositoriesGeneration\Configuration\Settings.cs" />
-    <Compile Include="RepositoriesGeneration\Core\CodeClassGeneratorСacheRepository.cs" />
-    <Compile Include="RepositoriesGeneration\Core\CodeClassGeneratorRepository.cs" />
+    <Compile Include="RepositoriesGeneration\Core\CacheRepositoryCodeGenerator.cs" />
+    <Compile Include="RepositoriesGeneration\Core\RepositoryCodeGenerator.cs" />
     <Compile Include="RepositoriesGeneration\Core\CodeClassGeneratorVersionedRepositoryService.cs" />
-    <Compile Include="RepositoriesGeneration\Core\CodeClassGeneratorVersionsRepository.cs" />
+    <Compile Include="RepositoriesGeneration\Core\SQL\ISqlScriptGenerator.cs" />
+    <Compile Include="RepositoriesGeneration\Core\VersionsRepositoryCodeGenerator.cs" />
     <Compile Include="RepositoriesGeneration\Core\RepositoryFactory.cs" />
     <Compile Include="DecoratorGeneration\Core\DecoratorCodeFactory.cs" />
     <Compile Include="RepositoriesGeneration\Core\SQL\SQLPostgresScriptGenerator.cs" />
     <Compile Include="RepositoriesGeneration\Core\SQL\SQLScriptGenerator.cs" />
-    <Compile Include="RepositoriesGeneration\Heplers\ExtensionMethods.cs" />
+    <Compile Include="RepositoriesGeneration\Enums\DatabaseType.cs" />
+    <Compile Include="RepositoriesGeneration\Helpers\ExtensionMethods.cs" />
     <Compile Include="Common\CodeHelper.cs" />
-    <Compile Include="RepositoriesGeneration\Heplers\PostgresColumnsHelper.cs" />
-    <Compile Include="RepositoriesGeneration\Heplers\SystemToPostgreSqlTypeMapper.cs" />
-    <Compile Include="RepositoriesGeneration\Heplers\RepositoryDataModelHelper.cs" />
-    <Compile Include="RepositoriesGeneration\Heplers\SyntaxAnalysisHelper.cs" />
-    <Compile Include="RepositoriesGeneration\Heplers\SystemToSqlTypeMapper.cs" />
+    <Compile Include="RepositoriesGeneration\Helpers\PostgresColumnsHelper.cs" />
+    <Compile Include="RepositoriesGeneration\Helpers\SystemToPostgreSqlTypeMapper.cs" />
+    <Compile Include="RepositoriesGeneration\Helpers\RepositoryDataModelHelper.cs" />
+    <Compile Include="RepositoriesGeneration\Helpers\SyntaxAnalysisHelper.cs" />
+    <Compile Include="RepositoriesGeneration\Helpers\SystemToSqlTypeMapper.cs" />
     <Compile Include="RepositoriesGeneration\Infrastructure\AnalysisError.cs" />
-    <Compile Include="RepositoriesGeneration\Infrastructure\BaseCodeClassGeneratorRepository.cs" />
+    <Compile Include="RepositoriesGeneration\Infrastructure\RepositoryCodeGeneratorAbstract.cs" />
     <Compile Include="Common\Infrastructure\ICodeClassGenerator.cs" />
     <Compile Include="RepositoriesGeneration\Infrastructure\RepositoryAnalysisError.cs" />
     <Compile Include="RepositoriesGeneration\Infrastructure\RepositoryFieldInfo.cs" />
@@ -205,6 +207,7 @@
     <Compile Include="RepositoriesGeneration\Services\RepositoryService.cs" />
     <Compile Include="RepositoriesGeneration\Yumapos.Infrastructure.Clone\Attributes\DataAccessAttribute.cs" />
     <Compile Include="RepositoriesGeneration\Yumapos.Infrastructure.Clone\Attributes\DataMany2ManyAttribute.cs" />
+    <Compile Include="RepositoriesGeneration\Yumapos.Infrastructure.Clone\Attributes\DataPostgresAttribute.cs" />
     <Compile Include="RepositoriesGeneration\Yumapos.Infrastructure.Clone\Attributes\DbIgnoreAttribute.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
